06-test-harness/test-runner.py
#!/usr/bin/env python3
import argparse
import json
import time
from datetime import datetime
from pathlib import Path

import yaml
from rich import print


def load_config(path: str) -> dict:
	with open(path, "r", encoding="utf-8") as f:
		return yaml.safe_load(f)


def run_scenario(scenario: dict, target: dict) -> dict:
	start = time.time()
	result = {
		"scenario": scenario.get("name"),
		"target": target.get("host"),
		"status": "passed",
		"details": {},
	}
	# Placeholder for actual interactions
	time.sleep(1)
	result["duration_s"] = round(time.time() - start, 3)
	return result


def main():
	parser = argparse.ArgumentParser()
	parser.add_argument("--config", default="test-config.yaml")
	args = parser.parse_args()

	config = load_config(args.config)
	results = []

	for target in config.get("targets", []):
		for scenario in config.get("scenarios", []):
			print(f"[bold cyan]Running[/] {scenario.get('name')} against {target.get('host')}...")
			res = run_scenario(scenario, target)
			results.append(res)

	out_dir = Path("results")
	out_dir.mkdir(parents=True, exist_ok=True)
	out_path = out_dir / f"test-results-{datetime.utcnow().strftime('%Y%m%dT%H%M%SZ')}.json"
	with open(out_path, "w", encoding="utf-8") as f:
		json.dump(results, f, indent=2)
	print(f"[green]Saved results to[/] {out_path}")


if __name__ == "__main__":
	main()
