06-test-harness/requirements.txt
pyyaml==6.0.2
requests==2.32.3
rich==13.8.0
